move greedy test out of recipe module tests
