+++ tests/PYME/localization/traveling_salesperson.py
@@ -17,3 +17,15 @@ def test_timeout_two_opt():
     assert elapsed - timeout < 1.25 * timeout
     assert best_distance < og_distance
     np.testing.assert_array_equal(np.sort(route), np.arange(positions.shape[0], dtype=int))
+
+def test_greedy():
+    from PYME.localization import traveling_salesperson as tsp
+    n = 500
+    x = np.random.rand(n) * 4e3
+    y = np.random.rand(n) * 4e3
+    positions = np.stack([x, y], axis=1)
+    start_ind = np.argmin(positions.sum(axis=1))
+    route, og_distance, final_distance = tsp.greedy_sort(positions, start_ind)
+    uni, counts = np.unique(route, return_counts=True)
+    np.testing.assert_array_equal(counts, 1)
+    assert final_distance < og_distance
